requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
azure-identity
functions-framework==3.*
google-cloud-storage
google-cloud-logging
flask-expects-json
python-dotenv
ciso8601
